refactor(arena): move the arena runtime out of the game hook
